+++ src/Recyclarr.TrashLib/ExceptionTypes/DuplicateInstancesException.cs
@@ -1,0 +1,11 @@
+namespace Recyclarr.TrashLib.ExceptionTypes;
+
+public class DuplicateInstancesException : Exception
+{
+    public IReadOnlyCollection<string> InstanceNames { get; }
+
+    public DuplicateInstancesException(IReadOnlyCollection<string> instanceNames)
+    {
+        InstanceNames = instanceNames;
+    }
+}